extract class GankAdapter which be used as base adapter
